Implement Size Limit on BlockingCollection to Reduce Memory Usage in Peak Pick Process
In the peak pick process, which employs the producer-consumer pattern for chromatogram analysis, we identified an issue with increasing memory usage due to an excessive number of chromatograms being prepared by the producer. To mitigate this, we have now implemented a size limit on the BlockingCollection, directly correlating it with the number of active consumers. This adjustment effectively controls the number of chromatograms queued at any given time, significantly reducing memory usage and enhancing the overall efficiency and stability of the system.