Improve peak pick process for lcms data (#290)
* Add Benchmark project

* WIP: try improving

* Implement Size Limit on BlockingCollection to Reduce Memory Usage in Peak Pick Process

In the peak pick process, which employs the producer-consumer pattern for chromatogram analysis, we identified an issue with increasing memory usage due to an excessive number of chromatograms being prepared by the producer. To mitigate this, we have now implemented a size limit on the BlockingCollection, directly correlating it with the number of active consumers. This adjustment effectively controls the number of chromatograms queued at any given time, significantly reducing memory usage and enhancing the overall efficiency and stability of the system.

* Optimize Chromatogram Data Management to Reduce GC Frequency

This update brings an optimization to the internal data management of chromatograms. By refining how data is handled and stored, we have successfully reduced the frequency of garbage collection (GC). This improvement in memory management enhances the current performance of our system and sets the foundation for more efficient processing, particularly in the context of future multi-threaded execution.

* Reduce Array Allocation by Utilizing a ArrayPool