Make ExecutableBar non-focusable by default
The `ExecutableBar` class in the `CompMs.Graphics.UI` namespace has been modified to include an additional line in its static constructor. Specifically, the `IsTabStopProperty` is now being overridden with a new `FrameworkPropertyMetadata` that sets its value to `false`. This change ensures that instances of `ExecutableBar` are not focusable via the tab key by default.