Add SpectraSimilarityMapViewModel with reactive properties
A new ViewModel class `SpectraSimilarityMapViewModel` has been added to the `CompMs.App.Msdial.ViewModel.Information` namespace. This class is designed to work with a model of type `SpectraSimilarityMapModel` and includes properties `MzBin`, `MzBegin`, `MzEnd`, and `Result`, which are synchronized with the model using ReactiveProperty. An `UpdateCommand` property of type `AsyncReactiveCommand` has been added, bound to the `UpdateSimilaritiesAsync` method of the model. The class inherits from `ViewModelBase` and uses Reactive Extensions for property synchronization and command handling. Necessary using directives have also been added.