Rename reverseDotProduct to sqreverseDotProduct
This commit updates the variable name `reverseDotProduct` to `sqreverseDotProduct` across multiple files in the `CompMs.Common.Algorithm.Scoring` and `CompMs.MsdialCore.Algorithm.Annotation` namespaces. This change reflects a new naming convention that emphasizes the squared nature of the dot product calculation. Additionally, comments have been added to the `GetReverseDotProduct` method to clarify that it represents the square of a typical dot product.