Updated Default Dot Product Thresholds (#591)
* Revert "Update dot product cutoff values across multiple files"

This reverts commit b30e08744885a18096aa6f99a6b5876a36854b09.

* Update notice for dot product cutoff value changes

Added a new `TextBlock` in the `HeaderedContentControl` to inform users about the changes in cutoff values for dot product calculations starting from version 5.5.250625. The previous text block was removed, and the new content explains the transition from squared values to native dot product values, along with instructions for adjusting project thresholds. This update also highlights the effect on the total score value in the MS-DIAL output.

* Update dot product cutoff parameters