Add SaveDataCommand for file saving functionality
Introduces the `SaveDataCommand` class in the `CompMs.Graphics.IO` namespace, implementing the `ICommand` interface. This class allows for saving data to a file using an asynchronous function, with support for file type filtering and a save file dialog. The `Execute` method manages the file stream, while `CanExecute` checks if a filter is set. Necessary using directives are also included.