Add CopyAsStringCommand for clipboard string operations
Introduces the `CopyAsStringCommand` class in the `CompMs.Graphics.IO` namespace. This class implements the `ICommand` interface to facilitate copying data to the clipboard as a string. It features a constructor that accepts an asynchronous `saveFunc`, a `Format` property defaulting to `DataFormats.Text`, and implementations for `CanExecute` and `Execute` methods. The `Execute` method utilizes a `MemoryStream` to store data, invokes the `saveFunc`, and sets the clipboard data using the UTF-8 encoded string from the memory stream.