Add lipidomics support with LipidmapsLinksModel
This commit introduces a new feature for lipidomics in the
`LcmsAlignmentModel` and `LcmsAnalysisModel` classes by adding
a `LipidmapsLinksModel` property. It initializes this model
when the target omics is set to lipidomics, utilizing a
`LipidmapsRestAPIHandler`. The `LipidmapsLinkViewModel` is
also integrated into the view models and XAML templates,
enabling the display of database links in the UI.
Additionally, the handling of peak detail view models is
refactored to conditionally include the new lipid maps links
model, improving functionality and user experience for
lipidomics data.