Refactor product ion abundance export functionality
Removed the `ExportProductIonAbundanceResultAsync` method from `LcmsAlignmentModel.cs` and replaced it with the `AdhocExportProductIonAbundanceResultCommand` in the view model. Updated key bindings in `LcmsMainView.xaml` to reflect this change, including the removal of the `Ctrl+Shift+B` binding. The `AdhocExportProductIonQuantifactionResultCommand` was also removed from `LcmsAlignmentViewModel.cs`, indicating a shift in how product ion abundance results are managed within the application.