Implement analysis pipeline and UI for TIMS-off imaging (#639)
* Update DisposableModelBase

* Update model classes for imaging

* Add IWholeImageResultModel interface

* Move SaveImagesModel from ImagingImms to Imaging directory

Move SaveImagesViewModel from ImagingImms to Imaging directory

* Add ImagingDims.WholeImageResultModel

* Add ImagingDims.ImagingDimsImageModel

* Update DimsProviderFactoryPrameter class

* Add ImagingDimsMethodModel class

* Add ImagingDims.WholeImageResultModel

* Update ImagingImmsImageViewModel

* Add ImagingDimsImageViewModel

* Update ImagingImmsMainViewModel

* Add ImagingDimsMainViewModel

* Update DatasetViewModel

* Add DimsImagingRibbon

* Add ImagingDims.ImagingMainView

Update ImagingDims.ImagingMainView

* Add DataTemplate for ImagingDimsMainViewModel

Introduced a new `DataTemplate` for the `ImagingDimsMainViewModel`, including necessary XML namespace declarations and an `ImagingMainView` element. The existing `DataTemplate` for `ImagingImmsMainViewModel` remains unchanged.

* Update MsdialDimsParameter for imaging support

Modified the MsdialDimsParameter constructor to include an
isImaging parameter, allowing for dynamic selection of
MachineCategory. Updated the CreateParameter method in
DatasetParameterSettingModel.cs to instantiate
MsdialDimsParameter with the appropriate isImaging value
based on the separation type.

* Add ImagingDimsMethodSettingModelFactory class

Introduce a new class `ImagingDimsMethodSettingModelFactory` to handle method settings for imaging dimensions (IDIMS). Update `MethodSettingModelFactory` to conditionally use this new factory based on the `MachineCategory` of `dimsStorage`. The new factory includes methods for creating various settings models tailored for imaging dimensions.

* Update Imaging CheckBox bindings and style triggers

Modified the "Imaging" CheckBox to bind the `IsEnabled` property to `DirectInfution.Value`. Removed the previous `MultiBinding` for `IsEnabled` and updated the style by eliminating a `DataTrigger` that set `IsChecked` to `False` based on `SeparationIonMobility.Value`.

* Add LipidmapsLink DataTemplate to ImagingMainView

Introduced a new `DataTemplate` for `LipidmapsLinkViewModel` in `ImagingMainView.xaml`. This includes a `TextBlock` with the text "Links", a tooltip for database links, and a vertical alignment set to stretch. An `info:LipidmapsLinkView` control was also added to the template.

* Refactor XAML bindings for clarity and consistency

Updated binding syntax in DataCollectionSettingView.xaml
to use explicit `Path` format for improved readability
and maintainability. Adjusted properties of `TextBox`,
`ContentControl`, and `StackPanel` to ensure uniformity
across the file without altering functionality.

* Enhance SaveAsync with cancellation and peak options

Updated the `SaveAsync` method in `IntensityImageModel` to accept `skipUnknownPeaks` and `token` parameters, improving its functionality and cancellation support. Adjusted calls to `SaveAsync` in `WholeImageResultModel` to utilize the new parameters, ensuring consistent behavior across the codebase.