Enhancement of fill percentage property accessibility (#659)
* Enhance FillPercentage property with getter and setter

Updated the `FillPercentage` property in `AlignmentSpotPropertyModel.cs` to include a setter. The setter checks for changes in value before updating `innerModel.FillParcentage` and raises the `OnPropertyChanged` event to notify listeners of the change.

* Make "Fill %" column editable in multiple views

Changed the `IsReadOnly` property for the "Fill %" column from `True` to `False` in several XAML files, enabling user editing of fill percentage values. Affected files include:
- DimsAlignmentSpotTableView.xaml
- GcmsAlignmentSpotTableView.xaml
- ImmsAlignmentSpotTableView.xaml
- LcimmsAlignmentSpotTableView.xaml
- LcmsAlignmentSpotTableView.xaml
- LcmsProteomicsAlignmentTableView.xaml